[Middleware_02] Fixed Speed int data type bug
If the time was less than an hour, the duration was 0, and the speed infinity...

+++ src/main/java/edu/oakland/helper/middleware02/TrackDataCalculator.java
@@ -26,7 +26,7 @@ public class TrackDataCalculator {
     LocalDateTime time1 = point1.getTime();
     LocalDateTime time2 = point2.getTime();
     Duration duration = Duration.between(time1, time2);
-    long newDuration = duration.getSeconds() / 3600;
+    double newDuration = duration.getSeconds() / 3600.0;
 
     float deltaY = longitude2 - longitude1;
     float deltaX = lattitude2 - lattitude1;
